First build that ends in a graph. Whether it is correct is a another question.

+++ JeuDeTaquin/Graph/GraphItem.h
@@ -1,7 +1,7 @@
 #pragma once
 
 struct GraphItem {
-    double X;
-    double Y;
+    float X;
+    int Y;
     double Avg;
 };
